backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
scikit-learn==1.5.1
numpy>=1.26.0
nltk==3.8.1
